Rename the faults module
